Reorganise shader functionality into src/shader
