+++ src/render/HCloudView.cpp
@@ -10,6 +10,7 @@
 #include "glutil.h"
 #include "QtLogger.h"
 #include "Shader.h"
+#include "ShaderProgram.h"
 #include "streampagecache.h"
 #include "util.h"
