+++ src/render/View3D.cpp
@@ -23,6 +23,7 @@
 #include "PointViewerMainWindow.h"
 #include "TriMesh.h"
 #include "Shader.h"
+#include "ShaderProgram.h"
 #include "tinyformat.h"
 #include "util.h"
